feat: create header component and refactor filenames
