+++ components/header.tsx
@@ -1,41 +1,29 @@
-import "../styles/App.scss";
-import Head from "next/head";
 import Link from "next/link";
 import { motion } from "framer-motion";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faGithub, faLinkedin } from "@fortawesome/free-brands-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-export default function Project({ children, pageTitle, description }) {
-  const page = {
-    hidden: {
-      opacity: 0,
-    },
-    pageShow: {
-      opacity: 1,
-    },
-  };
+type HeaderProps = {
+ logoLink: string;
+}
 
+const Header = ({logoLink}: HeaderProps) => {
   return (
-    <>
-      <Head>
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta charSet="utf-8" />
-        <meta name="Description" content={description}></meta>
-        <title>{pageTitle}</title>
-      </Head>
-      <nav className="project-nav">
+    <header>
+      <nav className="home-nav">
         <div className="space-between">
-          <Link href="/">
+          <Link href={logoLink}>
             <div className="logo">AG</div>
           </Link>
           <ul className="nav-list">
             <li>
               <motion.a
-                href="/Alexander_Grattan_Resume.pdf"
+                href="https://drive.google.com/file/d/1PgvpHThs5XjTwGZgib9ZTVLa8QbhulWp/view?usp=sharing"
                 target="_blank"
                 rel="noopener noreferrer"
                 whileHover={{ scale: 1.1 }}
                 whileTap={{ scale: 0.9 }}
+                title="Download Alexander's Resume"
               >
                 Resume
               </motion.a>
@@ -45,10 +33,12 @@ export default function Project({ children, pageTitle, description }) {
                 href="https://github.com/agrattan0820"
                 target="_blank"
                 rel="noopener noreferrer"
-                whileHover={{ scale: 1.1 }}
-                whileTap={{ scale: 0.9 }}
+                whileHover={{ scale: 1.05 }}
+                whileTap={{ scale: 0.95 }}
+                title="Go to Alexander's GitHub"
               >
                 <FontAwesomeIcon icon={faGithub} size="2x" />
+                <span className="header-hidden-text">GitHub</span>
               </motion.a>
             </li>
             <li>
@@ -56,16 +46,19 @@ export default function Project({ children, pageTitle, description }) {
                 href="https://www.linkedin.com/in/alexander-grattan/"
                 target="_blank"
                 rel="noopener noreferrer"
-                whileHover={{ scale: 1.1 }}
-                whileTap={{ scale: 0.9 }}
+                whileHover={{ scale: 1.05 }}
+                whileTap={{ scale: 0.95 }}
+                title="Connect with Alexander on LinkedIn"
               >
                 <FontAwesomeIcon icon={faLinkedin} size="2x" />
+                <span className="header-hidden-text">LinkedIn</span>
               </motion.a>
             </li>
           </ul>
         </div>
       </nav>
-      <main className="project-main">{children}</main>
-    </>
+    </header>
   );
-}
+};
+
+export default Header;
