+++ pages/index.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useState, useRef } from "react";
 
-import Head from "next/head";
 import { GetServerSideProps, NextPage } from "next";
-import IntroOverlay from "../components/introOverlay";
+import IntroOverlay from "../components/intro-overlay";
 import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/dist/ScrollTrigger";
 import {
@@ -19,6 +18,7 @@ import { motion } from "framer-motion";
 import Link from "next/link";
 import { projectsList } from "../utils/project-data";
 import SEO from "../components/seo";
+import Header from "../components/header";
 
 type HomepageProps = {
   project: string | false;
@@ -194,57 +194,7 @@ const Homepage: NextPage<HomepageProps> = ({ project }) => {
       />
       {animationComplete === false && <IntroOverlay />}
       <div className="after-animation">
-        <header>
-          <nav className="home-nav">
-            <div className="space-between">
-              <Link href="/">
-                <div className="logo">AG</div>
-              </Link>
-
-              <ul className="nav-list">
-                <li>
-                  <motion.a
-                    href="https://drive.google.com/file/d/1PgvpHThs5XjTwGZgib9ZTVLa8QbhulWp/view?usp=sharing"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    whileHover={{ scale: 1.1 }}
-                    whileTap={{ scale: 0.9 }}
-                    title="Download Alexander's Resume"
-                  >
-                    Resume
-                  </motion.a>
-                </li>
-                <li>
-                  <motion.a
-                    href="https://github.com/agrattan0820"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    whileHover={{ scale: 1.05 }}
-                    whileTap={{ scale: 0.95 }}
-                    title="Go to Alexander's GitHub"
-                  >
-                    <FontAwesomeIcon icon={faGithub} size="2x" />
-                    <span className="header-hidden-text">GitHub</span>
-                  </motion.a>
-                </li>
-                <li>
-                  <motion.a
-                    href="https://www.linkedin.com/in/alexander-grattan/"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    whileHover={{ scale: 1.05 }}
-                    whileTap={{ scale: 0.95 }}
-                    title="Connect with Alexander on LinkedIn"
-                  >
-                    <FontAwesomeIcon icon={faLinkedin} size="2x" />
-                    <span className="header-hidden-text">LinkedIn</span>
-                  </motion.a>
-                </li>
-              </ul>
-            </div>
-          </nav>
-        </header>
-
+      <Header logoLink="/" />
         <main className="main-home">
           <div className="cta">
             <h1 className="title">
